requirements.txt
Flask>=2.0
flask-cors
gunicorn
psycopg2-binary
python-dotenv
